backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1